Move refresh button state into its own function
