+++ app/src/main/java/com/capyreader/app/ui/PreferencesComposableExt.kt
@@ -6,4 +6,5 @@ import androidx.compose.runtime.collectAsState
 import com.jocmp.capy.preferences.Preference
 
 @Composable
-fun <T> Preference<T>.collectChangesWithDefault(): State<T> = changes().collectAsState(initial = defaultValue())
+fun <T> Preference<T>.collectChangesWithDefault(initial: T = defaultValue()): State<T> =
+    changes().collectAsState(initial = initial)
