+++ app/src/main/java/com/capyreader/app/ui/articles/ArticleLayout.kt
@@ -47,6 +47,7 @@ import com.capyreader.app.refresher.RefreshInterval
 import com.capyreader.app.ui.articles.detail.ArticleView
 import com.capyreader.app.ui.articles.detail.CapyPlaceholder
 import com.capyreader.app.ui.articles.detail.rememberArticlePagination
+import com.capyreader.app.ui.articles.feeds.FeedList
 import com.capyreader.app.ui.articles.list.ArticleListTopBar
 import com.capyreader.app.ui.articles.list.EmptyOnboardingView
 import com.capyreader.app.ui.articles.list.PullToNextFeedBox
